Make async_framework similar to step framework from podgroup test, so it will be easier to merge them in the future
