+++ test/integration/scheduler/preemption/deferred_resize_preemption_test.go
@@ -34,6 +34,7 @@ import (
 	"k8s.io/kubernetes/pkg/scheduler"
 	configtesting "k8s.io/kubernetes/pkg/scheduler/apis/config/testing"
 	st "k8s.io/kubernetes/pkg/scheduler/testing"
+	"k8s.io/kubernetes/test/integration/scheduler/preemption/asyncframework"
 	testutils "k8s.io/kubernetes/test/integration/util"
 	"k8s.io/utils/ptr"
 )
@@ -67,7 +68,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			existingPods: []*v1.Pod{
 				initPausePod(&testutils.PausePodConfig{
 					Name:     "victim-1",
-					Priority: &LowPriority,
+					Priority: &asyncframework.LowPriority,
 					Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 						v1.ResourceCPU:    *resource.NewMilliQuantity(200, resource.DecimalSI),
 						v1.ResourceMemory: *resource.NewQuantity(100, resource.DecimalSI)},
@@ -76,7 +77,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			},
 			preemptorConfig: &testutils.PausePodConfig{
 				Name:     "preemptor-pod",
-				Priority: &HighPriority,
+				Priority: &asyncframework.HighPriority,
 				Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 					v1.ResourceCPU:    *resource.NewMilliQuantity(300, resource.DecimalSI),
 					v1.ResourceMemory: *resource.NewQuantity(100, resource.DecimalSI)},
@@ -91,7 +92,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			existingPods: []*v1.Pod{
 				initPausePod(&testutils.PausePodConfig{
 					Name:     "victim-1",
-					Priority: &LowPriority,
+					Priority: &asyncframework.LowPriority,
 					Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 						v1.ResourceCPU:    *resource.NewMilliQuantity(100, resource.DecimalSI),
 						v1.ResourceMemory: *resource.NewQuantity(50, resource.DecimalSI)},
@@ -99,7 +100,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 				}),
 				initPausePod(&testutils.PausePodConfig{
 					Name:     "victim-2",
-					Priority: &LowPriority,
+					Priority: &asyncframework.LowPriority,
 					Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 						v1.ResourceCPU:    *resource.NewMilliQuantity(100, resource.DecimalSI),
 						v1.ResourceMemory: *resource.NewQuantity(50, resource.DecimalSI)},
@@ -108,7 +109,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			},
 			preemptorConfig: &testutils.PausePodConfig{
 				Name:     "preemptor-pod",
-				Priority: &HighPriority,
+				Priority: &asyncframework.HighPriority,
 				Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 					v1.ResourceCPU:    *resource.NewMilliQuantity(300, resource.DecimalSI),
 					v1.ResourceMemory: *resource.NewQuantity(100, resource.DecimalSI)},
@@ -123,7 +124,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			existingPods: []*v1.Pod{
 				initPausePod(&testutils.PausePodConfig{
 					Name:     "victim-1",
-					Priority: &LowPriority,
+					Priority: &asyncframework.LowPriority,
 					Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 						v1.ResourceCPU:    *resource.NewMilliQuantity(200, resource.DecimalSI),
 						v1.ResourceMemory: *resource.NewQuantity(100, resource.DecimalSI)},
@@ -132,7 +133,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			},
 			preemptorConfig: &testutils.PausePodConfig{
 				Name:             "preemptor-pod",
-				Priority:         &HighPriority,
+				Priority:         &asyncframework.HighPriority,
 				PreemptionPolicy: new(v1.PreemptNever),
 				Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 					v1.ResourceCPU:    *resource.NewMilliQuantity(300, resource.DecimalSI),
@@ -148,7 +149,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			existingPods:    nil, // fits immediately
 			preemptorConfig: &testutils.PausePodConfig{
 				Name:     "preemptor-pod",
-				Priority: &HighPriority,
+				Priority: &asyncframework.HighPriority,
 				Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 					v1.ResourceCPU:    *resource.NewMilliQuantity(300, resource.DecimalSI),
 					v1.ResourceMemory: *resource.NewQuantity(100, resource.DecimalSI)},
@@ -166,7 +167,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			existingPods: []*v1.Pod{
 				initPausePod(&testutils.PausePodConfig{
 					Name:     "victim-1",
-					Priority: &LowPriority,
+					Priority: &asyncframework.LowPriority,
 					Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 						v1.ResourceCPU:    *resource.NewMilliQuantity(200, resource.DecimalSI),
 						v1.ResourceMemory: *resource.NewQuantity(100, resource.DecimalSI)},
@@ -175,7 +176,7 @@ func TestDeferredResizePodPreemption(t *testing.T) {
 			},
 			preemptorConfig: &testutils.PausePodConfig{
 				Name:     "preemptor-pod",
-				Priority: &HighPriority,
+				Priority: &asyncframework.HighPriority,
 				Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 					v1.ResourceCPU:    *resource.NewMilliQuantity(300, resource.DecimalSI),
 					v1.ResourceMemory: *resource.NewQuantity(100, resource.DecimalSI)},
@@ -359,7 +360,7 @@ func setUpPreemptionTestWithContext(t *testing.T, testCtx *testutils.TestContext
 	other := initPausePod(&testutils.PausePodConfig{
 		Name:     fmt.Sprintf("other-pod-%d", idx),
 		NodeName: nodeName1,
-		Priority: &HighPriority,
+		Priority: &asyncframework.HighPriority,
 		Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 			v1.ResourceCPU: resource.MustParse(otherPodCPURequest)},
 		},
@@ -374,7 +375,7 @@ func setUpPreemptionTestWithContext(t *testing.T, testCtx *testutils.TestContext
 	irrelevant := initPausePod(&testutils.PausePodConfig{
 		Name:     fmt.Sprintf("irrelevant-pod-%d", idx),
 		NodeName: nodeName2,
-		Priority: &HighPriority,
+		Priority: &asyncframework.HighPriority,
 		Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 			v1.ResourceCPU: *resource.NewMilliQuantity(200, resource.DecimalSI)},
 		},
@@ -389,7 +390,7 @@ func setUpPreemptionTestWithContext(t *testing.T, testCtx *testutils.TestContext
 	pod := initPausePod(&testutils.PausePodConfig{
 		Name:     fmt.Sprintf("deferred-pod-%d", idx),
 		NodeName: nodeName1,
-		Priority: &LowPriority,
+		Priority: &asyncframework.LowPriority,
 		Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 			v1.ResourceCPU: resource.MustParse(deferredPodCPURequest)},
 		},
@@ -770,7 +771,7 @@ func TestDeferredResizeQueueingHandlers(t *testing.T) {
 		pod := initPausePod(&testutils.PausePodConfig{
 			Name:     "deferred-pod",
 			NodeName: "node1",
-			Priority: &LowPriority,
+			Priority: &asyncframework.LowPriority,
 			Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 				v1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI)},
 			},
@@ -867,7 +868,7 @@ func TestDeferredResizeQueueingHandlers(t *testing.T) {
 		pod := initPausePod(&testutils.PausePodConfig{
 			Name:     "deferred-pod",
 			NodeName: "node1",
-			Priority: &LowPriority,
+			Priority: &asyncframework.LowPriority,
 			Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 				v1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI)},
 			},
@@ -971,7 +972,7 @@ func TestDeferredResizeQueueingHandlers(t *testing.T) {
 		pod := initPausePod(&testutils.PausePodConfig{
 			Name:     "deferred-pod",
 			NodeName: "node1",
-			Priority: &LowPriority,
+			Priority: &asyncframework.LowPriority,
 			Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 				v1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI)},
 			},
@@ -1069,7 +1070,7 @@ func TestDeferredResizeQueueingHandlers(t *testing.T) {
 		pod := initPausePod(&testutils.PausePodConfig{
 			Name:     "deferred-pod-del",
 			NodeName: "node1",
-			Priority: &LowPriority,
+			Priority: &asyncframework.LowPriority,
 			Resources: &v1.ResourceRequirements{Requests: v1.ResourceList{
 				v1.ResourceCPU: *resource.NewMilliQuantity(100, resource.DecimalSI)},
 			},
